requirements.txt
livekit-agents[images]>=0.12.16,<1.0.0
livekit-plugins-google>=0.11.0
python-dotenv
livekit
asyncio
